Got a bit more working
Still not sure why adding in serial print lines make the controller work
but it does. Also, I have a suspicion that the smoothed values are
causing problems so i am outputting what the smoothed value is. Lastly,
i want to try again at calibrating the ESC's in hope that it will fix
the one motor being stronger problem.